Changed FastJets init1() method.
It uses a switch instead of a gnarly if then else sequence. Also had
to change order of enum in FastJets.hh, which could give unintended
bugs if someone uses that enum as ints instead of by name...

This is not a functional change, but a stylistic one that (hopefully)
aids in long-term code maintenance.  It doesn't need to be merged in
with the other fast-jets changes at the end of the day.